refactor(phase2): clarify analysis CLI navigation
