+++ scripts/02_phase2_analysis/plot_heatmap.py
@@ -1,147 +1,14 @@
 #!/usr/bin/env python3
 """Plot a configured Phase 2 sweep heatmap."""
 
-from __future__ import annotations
-
-import argparse
 from pathlib import Path
 import sys
 
-sys.path.insert(0, str(Path(__file__).parents[2] / "src"))
+sys.path.insert(0, str(Path(__file__).resolve().parents[2] / "src"))
 
-from sim_swim.analysis.cli_profiles import (
-    args_from_profile,
-    format_profile_description,
-    format_profile_listing,
-    key_value_args_to_cli_args,
-    list_profile_entries,
-    load_profile_entry,
-    split_config_key,
-    validate_profile_role,
-)
-from sim_swim.analysis.heatmaps import (
-    dt_star_torque,
-    generic_multi_run,
-    hook_overstretch,
-    local_scale_mode,
-    motor_scale_collapse,
-    shape_stability_grid,
-)
+from sim_swim.analysis.phase2_heatmap import HEATMAP_MAIN, main
 
-
-HEATMAP_MAIN = {
-    "motor_scale_collapse": motor_scale_collapse.main,
-    "dt_star_torque": dt_star_torque.main,
-    "local_scale_mode": local_scale_mode.main,
-    "shape_stability_grid": shape_stability_grid.main,
-    "hook_overstretch": hook_overstretch.main,
-    "generic_multi_run": generic_multi_run.main,
-}
-
-
-def _heatmap_entries(*, canonical_only: bool = False) -> list[dict[str, object]]:
-    entries = list_profile_entries(role="heatmap", canonical_only=canonical_only)
-    entries.extend(
-        entry
-        for entry in list_profile_entries(role="sweep", canonical_only=canonical_only)
-        if entry["kind"] == "generic_multi_run"
-    )
-    seen: set[str] = set()
-    unique_entries: list[dict[str, object]] = []
-    for entry in entries:
-        path = str(entry["path"])
-        if path in seen:
-            continue
-        seen.add(path)
-        unique_entries.append(entry)
-    return unique_entries
-
-
-def _has_option(args: list[str], option_name: str) -> bool:
-    return any(arg == option_name or arg.startswith(f"{option_name}=") for arg in args)
-
-
-def _first_option_value(args: list[str], option_name: str) -> str | None:
-    for index, arg in enumerate(args):
-        if arg == option_name and index + 1 < len(args):
-            return args[index + 1]
-        if arg.startswith(f"{option_name}="):
-            return arg.split("=", 1)[1]
-    return None
-
-
-def _with_run_dir_defaults(args: list[str]) -> list[str]:
-    run_dir = _first_option_value(args, "--run-dir")
-    if run_dir is not None and not _has_option(args, "--summary-csv"):
-        args = [*args, "--summary-csv", str(Path(run_dir) / "summary.csv")]
-    if _has_option(args, "--output-dir"):
-        return args
-    if run_dir is not None:
-        return [*args, "--output-dir", str(Path(run_dir) / "plots")]
-    summary_csv = _first_option_value(args, "--summary-csv")
-    if summary_csv is None:
-        return args
-    return [*args, "--output-dir", str(Path(summary_csv).parent / "plots")]
-
-
-def main(argv: list[str] | None = None) -> None:
-    raw_argv = list(sys.argv[1:] if argv is None else argv)
-    config_from_key, parser_argv = split_config_key(raw_argv)
-    parser_argv = key_value_args_to_cli_args(parser_argv)
-
-    parser = argparse.ArgumentParser(description=__doc__)
-    parser.add_argument(
-        "--config",
-        type=Path,
-        default=None,
-        help="Heatmap profile YAML under conf/phase2_sweeps/ or conf/phase2_multi_run/.",
-    )
-    parser.add_argument(
-        "--list-profiles",
-        action="store_true",
-        help="List available heatmap profiles and exit.",
-    )
-    parser.add_argument(
-        "--list-canonical-profiles",
-        action="store_true",
-        help="List canonical heatmap profiles and exit.",
-    )
-    parser.add_argument(
-        "--describe-profile",
-        action="store_true",
-        help="Print profile metadata for the selected config and exit.",
-    )
-    args, passthrough = parser.parse_known_args(parser_argv)
-    if config_from_key is not None and args.config is not None:
-        parser.error("Use either config=PATH or --config PATH (not both)")
-    if args.list_profiles or args.list_canonical_profiles:
-        entries = _heatmap_entries(canonical_only=args.list_canonical_profiles)
-        for line in format_profile_listing(entries):
-            print(line)
-        return
-    config = config_from_key or args.config
-    if config is None:
-        parser.error("config=PATH or --config PATH is required")
-
-    entry = load_profile_entry(config)
-    kind = entry["kind"]
-    if kind not in HEATMAP_MAIN:
-        choices = ", ".join(sorted(HEATMAP_MAIN))
-        raise SystemExit(f"Unknown heatmap kind {kind!r}. Expected one of: {choices}")
-    if args.describe_profile:
-        for line in format_profile_description(entry, _heatmap_entries()):
-            print(line)
-        return
-    if kind != "generic_multi_run":
-        validate_profile_role(entry, "heatmap")
-
-    if kind == "generic_multi_run":
-        effective_args = _with_run_dir_defaults(
-            ["--campaign-config", str(config)] + key_value_args_to_cli_args(passthrough)
-        )
-    else:
-        effective_args = _with_run_dir_defaults(args_from_profile(entry) + passthrough)
-    HEATMAP_MAIN[kind](effective_args)
+__all__ = ["HEATMAP_MAIN", "main"]
 
 
 if __name__ == "__main__":
