+++ src/sim_swim/analysis/phase2_158_diagnostics.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import argparse
 import csv
 from dataclasses import dataclass
 from datetime import datetime
@@ -1025,3 +1026,65 @@ def analyze_phase2_158_diagnostics(config: Phase2158DiagnosticConfig) -> Path:
     )
     logger.info("Wrote diagnostics to %s", config.output_dir)
     return config.output_dir
+
+
+def _parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(description=__doc__)
+    parser.add_argument(
+        "--campaign-config",
+        type=Path,
+        default=Path("conf/phase2_multi_run/flagella_count_duration_3s_r1.yaml"),
+        help="Phase 2 multi-run campaign config used to reconstruct condition topology.",
+    )
+    parser.add_argument(
+        "--input-dir",
+        type=Path,
+        default=Path("outputs/phase2_multi_run/flagella_count_duration_3s_r1"),
+        help="Existing raw multi-run output directory.",
+    )
+    parser.add_argument(
+        "--output-dir",
+        type=Path,
+        default=None,
+        help="Output directory. Defaults to outputs/YYYY-MM-DD/HHMMSS/...",
+    )
+    parser.add_argument(
+        "--fail-condition-id",
+        action="append",
+        default=None,
+        help="Condition ID to include in first-fail event plots. Can be repeated.",
+    )
+    parser.add_argument(
+        "--first-fail-window-s",
+        type=float,
+        default=0.25,
+        help="Seconds before/after first fail to include in event tables and plots.",
+    )
+    parser.add_argument("--overwrite", action="store_true")
+    parser.add_argument(
+        "overrides",
+        nargs="*",
+        help="Optional KEY=VALUE campaign overrides for analysis reproducibility.",
+    )
+    return parser.parse_args()
+
+
+def main() -> None:
+    args = _parse_args()
+    output_dir = (
+        args.output_dir if args.output_dir is not None else default_output_dir()
+    )
+    result = analyze_phase2_158_diagnostics(
+        Phase2158DiagnosticConfig(
+            campaign_config=args.campaign_config,
+            input_dir=args.input_dir,
+            output_dir=output_dir,
+            fail_condition_ids=tuple(
+                args.fail_condition_id or DEFAULT_FAIL_CONDITION_IDS
+            ),
+            first_fail_window_s=float(args.first_fail_window_s),
+            overwrite=bool(args.overwrite),
+            cli_overrides=tuple(args.overrides),
+        )
+    )
+    print(result)
